Refactor globalEnvironment
