+++ src/utils/environment/index.ts
@@ -5,8 +5,7 @@ export * from './Environment';
 export * from './constructEnvironment';
 export * from './ensureAccountAddress';
 export * from './getAccounts';
-export * from './getGlobalEnvironment';
 export * from './getWeb3Options';
+export * from './globalEnvironment';
 export * from './initGlobalEnvironment';
 export * from './initTestEnvironment';
-export * from './setGlobalEnvironment';
